Fix sudo on test package

+++ test_package/conanfile.py
@@ -10,7 +10,7 @@ class TestPackageConan(ConanFile):
     settings = "os", "compiler", "build_type", "arch"
     generators = "cmake"
 
-    def is_sudo_enabled():
+    def is_sudo_enabled(self):
         if "CONAN_SYSREQUIRES_SUDO" not in os.environ:
             if os.name == 'posix' and os.geteuid() == 0:
                 return False
@@ -21,21 +21,17 @@ class TestPackageConan(ConanFile):
     def copy_utc(self):
         if tools.os_info.is_linux:
             zoneinfo_dir = os.path.join(os.sep, "usr", "share", "zoneinfo")
-            try:
+            if not os.path.exists(os.path.join(zoneinfo_dir, "UTC")):
                 sudo = "sudo " if self.is_sudo_enabled() else ""
                 if not os.path.exists(zoneinfo_dir):
                     self.run("{} mkdir {}".format(sudo, zoneinfo_dir))
                 self.run("{} cp UTC {}".format(sudo, os.path.join(zoneinfo_dir, "UTC")))
-            except:
-                pass
         elif tools.os_info.is_macos:
             zoneinfo_dir = os.path.join(os.sep, "etc", "zoneinfo")
-            try:
+            if not os.path.exists(os.path.join(zoneinfo_dir, "UTC")):
                 if not os.path.exists(zoneinfo_dir):
                     self.run("mkdir {}".format(zoneinfo_dir))
                 self.run("cp UTC {}".format(os.path.join(zoneinfo_dir, "UTC")))
-            except:
-                pass
 
     def build(self):
         cmake = CMake(self)
